requirements.txt
boto3==1.34.131
pathspec==0.12.1
botocore==1.34.131
dvc[s3]==3.54.1
fastapi==0.115.12
httpx==0.28.1
joblib==1.4.2
numpy==1.26.4
pandas==2.0.1
Pillow==11.3.0
pydantic==2.12.5
pytest==8.3.2
python-multipart==0.0.20
scikit-learn==1.6.1
uvicorn==0.30.6
qdrant-client==1.9.0